Trimmed PDR
The document describing the process was verbose and at times not a fit for where PRAGMA is today and the near future. For many members, English is a second language and simplifying the document should make it easier for all. We can later revise the document as we and our needs grow.