Refined new projects pdr (3)
We can write a future pdr for the life cycle of a project that covers some of what I have removed as its not pertinent to onboarding new projects. 
I have also amended sections to accommodate for our statutes, the required recommendations and GA final approval. 

Per the board pdr on voting, the voting is done by show of hands or by equivalent electronic means; the vote can be "in favor" or "against" so we should be able to show our vote on this pdr in github